+++ src/private/release/InvokeNovaRepositoryPublishCommand.ps1
@@ -1,0 +1,8 @@
+function Invoke-NovaRepositoryPublishCommand {
+    [CmdletBinding()]
+    param(
+        [Parameter(Mandatory)][hashtable]$PublishParameters
+    )
+
+    Publish-PSResource @PublishParameters
+}